+++ club-website/src/components/announcement.tsx
@@ -6,7 +6,7 @@ interface AnnouncementProps {
 const Announcement: React.FC<AnnouncementProps> = ({ date, title }) => {
   return (
     <a>
-      <div className="bg-[#F6F6F6] p-4 rounded-xl hover:translate-y-2 duration-200">
+      <div className="bg-[#F6F6F6] p-4 rounded-xl hover:-translate-y-2 duration-200">
         <p>{date}</p>
         <p>{title}</p>
       </div>
