+++ club-website/src/components/footer.tsx
@@ -43,7 +43,7 @@ const LINKS = [
 
 export default function Footer() {
   return (
-    <main className="items-start justify-between lg:text-base md:text-md text-sm grid grid-rows-3 md:grid-cols-5 grid-cols-2 gap-8 lg:p-24 lg:ps-32 lg:pe-32 md:p-16 md:ps-16 md:pe-16 p-8 ps-12 pe-12">
+    <footer className="items-start justify-between lg:text-base md:text-md text-sm grid grid-rows-2 md:grid-cols-5 grid-cols-2 gap-8 lg:p-24 lg:ps-32 lg:pe-32 md:p-16 md:ps-16 md:pe-16 p-8 ps-12 pe-12">
       <div
         className="row-start-3 md:row-start-1 col-start-1 col-span-2"
         id="logo"
@@ -178,6 +178,6 @@ export default function Footer() {
         </a>
         <p>-</p>
       </div>
-    </main>
+    </footer>
   );
 }
